Add MusicIndexServiceUtils
To prevent MusicIndexService from directly relying on MediaFileJPSupport
and JpsonicComparators.
